feat: поправлены мутации работы со значением
