+++ client/gql/dcis/mutations/sheet/change_value.graphql
@@ -1,4 +1,4 @@
-mutation ChangeValue($documentId: ID!, $sheetId: Int!, $columnId: Int!, $rowId: Int!, $value: String!) {
+mutation ChangeValue($documentId: ID!, $sheetId: ID!, $columnId: ID!, $rowId: ID!, $value: String!) {
   changeValue(input: {
     documentId: $documentId,
     sheetId: $sheetId,
@@ -7,5 +7,7 @@ mutation ChangeValue($documentId: ID!, $sheetId: Int!, $columnId: Int!, $rowId: 
     value: $value
   }) {
     success
+    value
+    updatedAt
   }
 }
