+++ server/apps/dcis/schema/mutations/sheet_mutations.py
@@ -22,11 +22,9 @@ from apps.dcis.services.sheet_services import (
     change_cells_option,
     change_column_dimension,
     change_row_dimension,
-    move_merged_cells,
-)
-from apps.dcis.services.value_services import (
     create_file_value_archive,
     get_file_value_files,
+    move_merged_cells,
     update_or_create_file_value,
     update_or_create_value,
 )
@@ -228,29 +226,18 @@ class ChangeCellsOptionMutation(BaseMutation):
                 return ChangeCellsOptionMutation(changed_options=change_cells_option(cells, field, value))
 
 
-class UnloadFileValueArchiveMutation(BaseMutation):
-    """Выгрузка архива значения ячейки типа `Файл`."""
-
-    class Input:
-        value_id = graphene.ID(required=True, description='Идентификатор значения ячейки')
-
-    src = graphene.String(description='Ссылка на сгенерированный архив')
-
-    @staticmethod
-    @permission_classes((IsAuthenticated,))
-    def mutate_and_get_payload(root: None, info: ResolveInfo, value_id: str):
-        return UnloadFileValueArchiveMutation(src=create_file_value_archive(get_object_or_404(Value, pk=value_id)))
-
-
 class ChangeValueMutation(BaseMutation):
     """Изменение значения ячейки."""
 
     class Input:
         document_id = graphene.ID(required=True, description='Идентификатор документа')
-        sheet_id = graphene.Int(required=True, description='Идентификатор листа')
-        column_id = graphene.Int(required=True, description='Идентификатор колонки')
-        row_id = graphene.Int(required=True, description='Идентификатор строки')
+        sheet_id = graphene.ID(required=True, description='Идентификатор листа')
+        column_id = graphene.ID(required=True, description='Идентификатор колонки')
+        row_id = graphene.ID(required=True, description='Идентификатор строки')
         value = graphene.String(required=True, description='Значение')
+
+    value = graphene.String(required=True, description='Измененное значение')
+    updated_at = graphene.DateTime(required=True, description='Дата изменения')
 
     @staticmethod
     @permission_classes((IsAuthenticated,))
@@ -265,14 +252,14 @@ class ChangeValueMutation(BaseMutation):
     ):
         document: Document = get_object_or_404(Document, pk=from_global_id(document_id)[1])
         sheet: Sheet = get_object_or_404(Sheet, pk=sheet_id)
-        val, _ = update_or_create_value(
+        result = update_or_create_value(
             document=document,
             sheet=sheet,
             column_id=column_id,
             row_id=row_id,
             value=value
         )
-        return ChangeValueMutation(value=val)
+        return ChangeValueMutation(value=result.value.value, updated_at=result.updated_at)
 
 
 class ChangeFileValueMutation(BaseMutation):
@@ -280,13 +267,15 @@ class ChangeFileValueMutation(BaseMutation):
 
     class Input:
         document_id = graphene.ID(required=True, description='Идентификатор документа')
-        sheet_id = graphene.Int(required=True, description='Идентификатор листа')
-        column_id = graphene.Int(required=True, description='Идентификатор колонки')
-        row_id = graphene.Int(required=True, description='Идентификатор строки')
+        sheet_id = graphene.ID(required=True, description='Идентификатор листа')
+        column_id = graphene.ID(required=True, description='Идентификатор колонки')
+        row_id = graphene.ID(required=True, description='Идентификатор строки')
         value = graphene.String(required=True, description='Значение')
         remaining_files = graphene.List(graphene.NonNull(graphene.ID), required=True, description='Оставшиеся файлы')
         new_files = graphene.List(graphene.NonNull(Upload), required=True, description='Новые файлы')
 
+    value = graphene.String(required=True, description='Измененное значение')
+    updated_at = graphene.DateTime(required=True, description='Дата изменения')
     value_files = graphene.List(FileType, description='Измененные файлы')
 
     @staticmethod
@@ -304,7 +293,7 @@ class ChangeFileValueMutation(BaseMutation):
     ):
         document: Document = get_object_or_404(Document, pk=from_global_id(document_id)[1])
         sheet: Sheet = get_object_or_404(Sheet, pk=sheet_id)
-        val, _ = update_or_create_file_value(
+        result = update_or_create_file_value(
             user=info.context.user,
             document=document,
             sheet=sheet,
@@ -314,7 +303,25 @@ class ChangeFileValueMutation(BaseMutation):
             remaining_files=[int(from_global_id(global_id)[1]) for global_id in remaining_files],
             new_files=new_files
         )
-        return ChangeFileValueMutation(value=val, value_files=get_file_value_files(val))
+        return ChangeFileValueMutation(
+            value=result.value.value,
+            updated_at=result.updated_at,
+            value_files=get_file_value_files(result.value)
+        )
+
+
+class UnloadFileValueArchiveMutation(BaseMutation):
+    """Выгрузка архива значения ячейки типа `Файл`."""
+
+    class Input:
+        value_id = graphene.ID(required=True, description='Идентификатор значения ячейки')
+
+    src = graphene.String(description='Ссылка на сгенерированный архив')
+
+    @staticmethod
+    @permission_classes((IsAuthenticated,))
+    def mutate_and_get_payload(root: None, info: ResolveInfo, value_id: str):
+        return UnloadFileValueArchiveMutation(src=create_file_value_archive(get_object_or_404(Value, pk=value_id)))
 
 
 class SheetMutations(graphene.ObjectType):
@@ -328,12 +335,12 @@ class SheetMutations(graphene.ObjectType):
 
     change_cells_option = ChangeCellsOptionMutation.Field(required=True, description='Изменения опций ячейки')
 
-    unload_file_value_archive = UnloadFileValueArchiveMutation.Field(
-        required=True,
-        description='Выгрузка архива значения ячейки типа `Файл`'
-    )
     change_value = ChangeValueMutation.Field(required=True, description='Изменение значения ячейки')
     change_file_value = ChangeFileValueMutation.Field(
         required=True,
         description='Изменение значения ячейки типа `Файл`'
+    )
+    unload_file_value_archive = UnloadFileValueArchiveMutation.Field(
+        required=True,
+        description='Выгрузка архива значения ячейки типа `Файл`'
     )
